train/requirements.txt
google.cloud.storage
tensorflow==1.14.0
keras
pandas
